change ids to u32 instead of i32

+++ src-tauri/src/parser/encounter_state.rs
@@ -29,7 +29,7 @@ pub struct EncounterState {
 
     damage_log: HashMap<String, Vec<(i64, i64)>>,
     identity_log: HashMap<String, IdentityLog>,
-    cast_log: HashMap<String, HashMap<i32, Vec<i32>>>,
+    cast_log: HashMap<String, HashMap<u32, Vec<i32>>>,
 
     boss_hp_log: HashMap<String, Vec<BossHpLog>>,
 
@@ -294,7 +294,7 @@ impl EncounterState {
     pub fn on_skill_start(
         &mut self,
         source_entity: Entity,
-        skill_id: i32,
+        skill_id: u32,
         tripod_index: Option<TripodIndex>,
         tripod_level: Option<TripodLevel>,
         timestamp: i64,
@@ -395,8 +395,8 @@ impl EncounterState {
         proj_entity: &Entity,
         dmg_target_entity: &Entity,
         damage: i64,
-        skill_id: i32,
-        skill_effect_id: i32,
+        skill_id: u32,
+        skill_effect_id: u32,
         modifier: i32,
         target_current_hp: i64,
         target_max_hp: i64,
@@ -445,7 +445,7 @@ impl EncounterState {
         if proj_entity.entity_type == EntityType::PROJECTILE
             && is_battle_item(skill_effect_id, "attack")
         {
-            skill_effect_id = proj_entity.skill_effect_id as i32;
+            skill_effect_id = proj_entity.skill_effect_id;
         }
 
         let mut source_entity = self
@@ -589,7 +589,7 @@ impl EncounterState {
             let mut is_debuffed_by_support = false;
             let se_on_source = se_on_source
                 .iter()
-                .map(|(se, _)| (*se) as i32)
+                .map(|(se, _)| *se)
                 .collect::<Vec<_>>();
             for buff_id in se_on_source.iter() {
                 if !self
@@ -638,7 +638,7 @@ impl EncounterState {
             }
             let se_on_target = se_on_target
                 .iter()
-                .map(|(se, _)| (*se) as i32)
+                .map(|(se, _)| *se)
                 .collect::<Vec<_>>();
             for debuff_id in se_on_target.iter() {
                 if !self
@@ -974,7 +974,7 @@ fn is_support_class_id(class_id: u32) -> bool {
     class_id == 105 || class_id == 204 || class_id == 602
 }
 
-fn is_battle_item(skill_effect_id: i32, _item_type: &str) -> bool {
+fn is_battle_item(skill_effect_id: u32, _item_type: &str) -> bool {
     if let Some(item) = SKILL_EFFECT_DATA.get(&skill_effect_id) {
         if let Some(category) = item.item_category.as_ref() {
             return category == "useup_battle_item_common_attack";
@@ -983,7 +983,7 @@ fn is_battle_item(skill_effect_id: i32, _item_type: &str) -> bool {
     false
 }
 
-fn get_status_effect_data(buff_id: i32) -> Option<StatusEffect> {
+fn get_status_effect_data(buff_id: u32) -> Option<StatusEffect> {
     let buff = SKILL_BUFF_DATA.get(&buff_id);
     if buff.is_none() || buff.unwrap().icon_show_type == "none" {
         return None;
@@ -1239,8 +1239,8 @@ fn get_status_effect_buff_type_flags(buff: &SkillBuffData) -> u32 {
 }
 
 fn get_skill_name_and_icon(
-    skill_id: &i32,
-    skill_effect_id: &i32,
+    skill_id: &u32,
+    skill_effect_id: &u32,
     skill_name: String,
 ) -> (String, String) {
     if (*skill_id == 0) && (*skill_effect_id == 0) {
@@ -1289,7 +1289,7 @@ fn get_skill_name_and_icon(
     }
 }
 
-fn get_skill_name(skill_id: &i32) -> String {
+fn get_skill_name(skill_id: &u32) -> String {
     SKILL_DATA
         .get(skill_id)
         .map_or("".to_string(), |skill| skill.name.clone())
@@ -1302,7 +1302,7 @@ fn insert_data(
     prev_stagger: i32,
     damage_log: HashMap<String, Vec<(i64, i64)>>,
     identity_log: HashMap<String, IdentityLog>,
-    cast_log: HashMap<String, HashMap<i32, Vec<i32>>>,
+    cast_log: HashMap<String, HashMap<u32, Vec<i32>>>,
     boss_hp_log: HashMap<String, Vec<BossHpLog>>,
     stagger_log: Vec<(i32, f32)>,
     mut stagger_intervals: Vec<(i32, i32)>,
